feat: accumulate reads while waiting, then average
Instead of waiting the next tick doing nothing, we can read sensor
values and average the results from the whole sampling period.
This really makes a "sampling period" instead of a "sampling
interval". It will greately reduce noise on the physical sensor,
but also lower the response time of the internal "calibrated"
sensor from that of the physical sensor (few µs) to that of the
sampling time (few ms). I should later compare the results with
both conditions and decide on the best.