feat: lower Baud rate, make TX robust for poor USB
Arduino transmits 1 byte per sensor + 1 byte. With a high estimate
of 7 sensors (8 bytes to transmit), it takes 6 ms to transmit data
at a baud rate of 9600 bit/s. This leaves enough time for the
remainings of the 25 ms period, both on Arduino and on PC, to
measure and to plot.